Scheme.Parser

Description

Documentation

symbol :: Parser Char

Define a parser that recognizes allowed symbols in Scheme

readExpr :: String -> ThrowsError LispVal

Define a function to call parser and handle errors

readExprList :: String -> ThrowsError [LispVal]

spaces :: Parser ()

Ignore spaces when parsing input

data LispVal

A datatype that can hold any Lisp value

Constructors

Atom String

List [LispVal]

DottedList [LispVal] LispVal

Number Integer

String String

Bool Bool

PrimitiveFunc ([LispVal] -> ThrowsError LispVal)

Func

Fields

params :: [String]

vararg :: Maybe String

body :: [LispVal]

closure :: Env

IOFunc ([LispVal] -> IOThrowsError LispVal)

Port Handle

parseString :: Parser LispVal

Function for string parsing

parseAtom :: Parser LispVal

Function for parsing an atom, which is is a letter or symbol, followed by any number of letters, digits, or symbols.

parseNumber :: Parser LispVal

Function for parsing a number

parseList :: Parser LispVal

Function for parsing a list

parseDottedList :: Parser LispVal

Function for parsing a dotted list

parseQuoted :: Parser LispVal

Function for parsing a single-quoted value

parseExpr :: Parser LispVal

Function for parsing an expression, which can be am atom, string, number, quoted, list, or dotted list.

type Env = IORef [(String, IORef LispVal)]

nullEnv :: IO Env

type IOThrowsError = ExceptT LispError IO

data LispError

Define datatype to represent an error

Constructors

NumArgs Integer [LispVal]

TypeMismatch String LispVal

Parser ParseError

BadSpecialForm String LispVal

NotFunction String String

UnboundVar String String

Default String

showError :: LispError -> String

Make an instance of Show, and define how to print out error types

type ThrowsError = Either LispError

Type to represent functions that may throw an error or return a value

trapError :: (Show a, MonadError a m) => m String -> m String

Helper function to convert error to string representation

extractValue :: ThrowsError a -> a

Extract error value

readOrThrow :: Parser a -> String -> ThrowsError a

getVar :: Env -> String -> IOThrowsError LispVal

setVar :: Env -> String -> LispVal -> IOThrowsError LispVal

defineVar :: Env -> String -> LispVal -> IOThrowsError LispVal

makeFunc :: Monad m => Maybe String -> Env -> [LispVal] -> [LispVal] -> m LispVal

makeNormalFunc :: Env -> [LispVal] -> [LispVal] -> ExceptT LispError IO LispVal

makeVarArgs :: LispVal -> Env -> [LispVal] -> [LispVal] -> ExceptT LispError IO LispVal

load :: String -> IOThrowsError [LispVal]

apply :: LispVal -> [LispVal] -> IOThrowsError LispVal

Map eval function over the arguments

Define apply function

isBound :: Env -> String -> IO Bool

bindVars :: Env -> [(String, LispVal)] -> IO Env

liftThrows :: ThrowsError a -> IOThrowsError a

eval :: Env -> LispVal -> IOThrowsError LispVal

Create the evaluator. This maps a code datatype to a 'data' datatype.

showVal :: LispVal -> String

Function to specify how to display possible LispVals

unwordsList :: [LispVal] -> String

Helper function to glue a list of words together with spaces