Scheme.REPL

Description

Documentation

flushStr :: String -> IO ()

Immediately flush the stream

readPrompt :: String -> IO String

Print a prompt and read a line of input

evalAndPrint :: Env -> String -> IO ()

Evaluate string and trap errors

evalString :: Env -> String -> IO String

Evaluate string and print result

until_ :: Monad m => (a -> Bool) -> m a -> (a -> m ()) -> m ()

runOne :: [String] -> IO ()

Read input, call function, and print output in an infinite loop

runRepl :: IO ()

Run the read-evaluate-print loop